Holzinger and Swineford (1939) Mental Abilities Dataset

A classic dataset containing mental ability test scores from seventh- and eighth-grade children from two different schools (Pasteur and Grant-White). This dataset is widely used in psychometrics to demonstrate confirmatory factor analysis (CFA) and structural equation modeling (SEM).

Usage

HS_data

Format

A data frame with 301 rows and 9 variables:

x1

Visual perception test score (Visual factor).

x2

Cubes test score (Visual factor).

x3

Lozenges test score (Visual factor).

x4

Paragraph comprehension test score (Textual factor).

x5

Sentence completion test score (Textual factor).

x6

Word meaning test score (Textual factor).

x7

Speeded addition test score (Speed factor).

x8

Speeded counting of dots test score (Speed factor).

x9

Speeded discrimination of straight and curved capitals test score (Speed factor).

Source

Adapted from the lavaan package.

References

Holzinger, K. J., & Swineford, F. (1939). A study in factor analysis: The stability of a bi-factor solution. Supplementary Educational Monographs, 48. University of Chicago.

Rosseel, Y. (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36.